api/requirements.txt
fastapi
uvicorn
azure-mgmt-containerinstance
azure-identity